requirements.txt
fastapi
uvicorn
tensorflow
numpy
pydantic
requests
opencv-python
Pillow
python-multipart
scikit-learn
joblib
pandas